Extract editor dock manager
